src/backend/expungeservice/models/charge_types/severe_charge.py
from dataclasses import dataclass

from expungeservice.models.charge import ChargeType
from expungeservice.models.charge import ChargeUtil
from expungeservice.models.expungement_result import TypeEligibility, EligibilityStatus


@dataclass(frozen=True)
class SevereCharge(ChargeType):
    type_name: str = "Severe Charge"
    expungement_rules: str = (
        """Charges that are harsher than Class A Felonies in severity: namely, murder and treason."""
    )

    def type_eligibility(self, disposition):
        if ChargeUtil.dismissed(disposition):
            raise ValueError("Dismissed criminal charges should have been caught by another class.")
        elif ChargeUtil.convicted(disposition):
            return TypeEligibility(EligibilityStatus.INELIGIBLE, reason="Ineligible by omission from statute")
